move types package out of internal
This enables users to implement interface methods like CreateChannelMonitor()
