+++ channels.go
@@ -6,8 +6,8 @@ import (
 	"sync"
 
 	"github.com/quentinmit/go-pvaccess/internal/ctxlog"
-	"github.com/quentinmit/go-pvaccess/internal/server/types"
 	"github.com/quentinmit/go-pvaccess/pvdata"
+	"github.com/quentinmit/go-pvaccess/types"
 	"golang.org/x/sync/errgroup"
 )
