+++ internal/search/udp.go
@@ -12,9 +12,9 @@ import (
 	"github.com/quentinmit/go-pvaccess/internal/connection"
 	"github.com/quentinmit/go-pvaccess/internal/ctxlog"
 	"github.com/quentinmit/go-pvaccess/internal/proto"
-	"github.com/quentinmit/go-pvaccess/internal/server/types"
 	"github.com/quentinmit/go-pvaccess/internal/udpconn"
 	"github.com/quentinmit/go-pvaccess/pvdata"
+	"github.com/quentinmit/go-pvaccess/types"
 )
 
 const startupInterval = time.Second
