+++ internal/server/status/status.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 
 	"github.com/quentinmit/go-pvaccess/internal/ctxlog"
-	"github.com/quentinmit/go-pvaccess/internal/server/types"
 	"github.com/quentinmit/go-pvaccess/pvdata"
+	"github.com/quentinmit/go-pvaccess/types"
 )
 
 type ChannelProviderser interface {
